fixed dentry_update; add inode tracker

+++ test/dentry_read.sh
@@ -3,24 +3,35 @@
 export TEVFS_WORKSPACE=$(dirname "$0")
 cd $TEVFS_WORKSPACE
 
-# add a dentry
-echo -e "=== running dentry_add.sh ==="
-./dentry_add.sh --no-unmount
-echo -e "\n=== dentry_add.sh exit $? ==="
-
 source img-var.sh
-cino_num=59
+
+echo "$TEVFS_SIZE $TEVFS_NUM_INODES $TEVFS_IMAGEPATH $TEVFS_MOUNTPT"
+rm -f $TEVFS_IMAGEPATH
+
+truncate -s $TEVFS_SIZE $TEVFS_IMAGEPATH
+
+if ! sudo -E ./img-mkfs.sh; then
+    exit 1
+fi
+
+if ! sudo -E ./img-mount.sh; then
+    exit 1
+fi
 
 dname="parent"
 fname="child"
 dir=$TEVFS_MOUNTPT/"$dname"
 file=$dir/"$fname"
 
+mkdir $dir
+touch $file
+
 sudo sync
 
 res=$(stat "$dir")
 pino_num=$(echo -e "$res" | head -n3 | tail -n1 | awk '{print $4}')
-echo -e "\nstat $dir\n$res"
+res=$(stat "$file")
+cino_num=$(echo -e "$res" | head -n3 | tail -n1 | awk '{print $4}')
 
 rm build/dentry_read.x
 make build/dentry_read.x &> /dev/null
